% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_master.R
\name{cluster_master}
\alias{cluster_master}
\title{Gather SE Cluster from Regression Results for Output}
\usage{
cluster_master(regs, cluster_names)
}
\arguments{
\item{regs}{stored regression output in a list}

\item{cluster_names}{vector of cluster variable names}
}
\description{
Helper function to construct clustering counts
}
\keyword{internal}
