% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_to_markdown.R
\name{tex_to_markdown}
\alias{tex_to_markdown}
\title{Convert the TeX output to Markdown Table for Printing to Console}
\usage{
tex_to_markdown(out_table)
}
\arguments{
\item{out_table}{the latex output from `estout`}
}
\value{
out_table The LaTeX code for the table
}
\description{
Like stata's `estout`, output regression results nicely.
}
\keyword{internal}
