% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_titles.R
\name{textablr_column_names_for_tex}
\alias{textablr_column_names_for_tex}
\title{Tex Table Column Toppers}
\usage{
textablr_column_names_for_tex(column_titles, column_widths, copy = TRUE)
}
\arguments{
\item{column_titles}{names of the regression columns}

\item{column_widths}{how many columns in each name category?}

\item{copy}{copy the output to clipboard? makes it easy to paste into tex document}
}
\description{
Helper function to construct tex table column names
}
\examples{

# Suppose you have 6 regressions,
# 3 on the male subsample and 3 on the female subsample

c_title <- c("Male", "Female")
c_widths <- c(3, 3)

# textablr_column_names_for_tex(column_titles = c_title, column_widths = c_widths)

}
