% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_column.R
\name{textablr_diff_col}
\alias{textablr_diff_col}
\title{Helper Function for Summary Statistics as a LaTeX Table}
\usage{
textablr_diff_col(data, var_labels, group_var, group1, group2)
}
\arguments{
\item{data}{the data}

\item{var_labels}{named vector of variable labels.
This is required
For example, to get summaries of the `educ` variable `Education` and the `expr` variable `Experience`
make this `c("Education" = "educ", "Experience" = "expr")`
Note: this also sets the order of variables in the table}

\item{group_var}{the variable (if any) to group on}

\item{group1}{the first group, the difference will be group1 - group2}

\item{group2}{the second group, the difference will be group1 - group2}
}
\description{
Generate a column of differences and standard errors
}
\examples{

mtcars_var_labels <- c("MPG" = "mpg", "Displacement" = "disp", "Horse Power (?)" = "hp")
textablr_diff_col(mtcars, var_labels = mtcars_var_labels, "gear", 3, 5)

}
