% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_to_clipboard.R
\name{textablr_get_parameters}
\alias{textablr_get_parameters}
\alias{textablr_get_sumstat_names}
\alias{textablr_get_var_labels}
\alias{textablr_get_var_indicates}
\alias{textablr_get_cluster_labels}
\title{Tex Table Get Tibbles in Clipboard}
\usage{
textablr_get_sumstat_names()

textablr_get_var_labels()

textablr_get_var_indicates()

textablr_get_cluster_labels()
}
\description{
Helper function to load the tibbles for defining summary statistics, variable labels, etc
    into the clipboard, so they are easy to change.
}
