% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{.}
\alias{.}
\title{Globals to trick compile complaints about things that are undefined}
\format{An object of class \code{NULL} of length 0.}
\usage{
.
}
\arguments{
\item{NULL}{}
}
\description{
Helper function to define global variables
}
\keyword{internal}
