% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\docType{data}
\name{sumstat_names}
\alias{sumstat_names}
\alias{sumstat_names_rdrobust}
\alias{sumstat_include_default}
\alias{indicator_levels}
\alias{star_level_default}
\title{A few simple lookup vectors and tibbles to help with regression output}
\format{An object of class \code{data.frame} with 6 rows and 3 columns.}
\usage{
sumstat_names

sumstat_names_rdrobust

sumstat_include_default

indicator_levels

star_level_default
}
\arguments{
\item{NULL}{}
}
\description{
Helper function to define various lookups. If you are curious what is in these lookups or defaults, just use
    `textablr:::LOOKUP` to see. Eventually, I'll learn how to put the output in this help file...
}
\keyword{datasets}
\keyword{internal}
