% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_column.R
\name{textablr_mean_col}
\alias{textablr_mean_col}
\title{Helper Function for Summary Statistics as a LaTeX Table}
\usage{
textablr_mean_col(data, var_labels, group_var = NULL,
  group_condition = NULL)
}
\arguments{
\item{data}{the data}

\item{var_labels}{named vector of variable labels.
This is required
For example, to get summaries of the `educ` variable `Education` and the `expr` variable `Experience`
make this `c("Education" = "educ", "Experience" = "expr")`
Note: this also sets the order of variables in the table}

\item{group_var}{the variable (if any) to group on}

\item{group_condition}{the filter on the group_var}
}
\description{
Generate a column of means and standard deviations
}
\examples{

mtcars_var_labels <- c("MPG" = "mpg", "Displacement" = "disp", "Horse Power (?)" = "hp")
textablr_mean_col(mtcars, var_labels = mtcars_var_labels, group_var = "cyl", group_condition = 4)

}
