% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_means_table.R
\name{textablr_means}
\alias{textablr_means}
\title{Output and Compute Summary Statistics as a LaTeX Table}
\usage{
textablr_means(..., file = "")
}
\arguments{
\item{...}{columns of summary statistics (means from `textablr_mean_col` and differences from `textablr_diff_col`)}

\item{file}{output location, if blank only prints to viewer with `gt()`}
}
\description{
Output summary stats results nicely.
}
\examples{

# simple table
# means of mtcars for gear 3 and 4 and the difference

mtcars_var_labels <- c("MPG" = "mpg", "Displacement" = "disp", "Horse Power (?)" = "hp")

col1 <- textablr_mean_col(mtcars, var_labels = mtcars_var_labels,
  group_var = "gear", group_condition = "3")

col2 <- textablr_mean_col(mtcars, var_labels = mtcars_var_labels,
  group_var = "gear", group_condition = "4")

col3 <- textablr_diff_col(mtcars, var_labels = mtcars_var_labels,
  group_var = "gear", group1 = "3", group2 = "4")

textablr_means(col1, col2, col3)

}
