% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_fe_master.R
\name{x_fe_master}
\alias{x_fe_master}
\title{Compile Regression Coefficients for Output}
\usage{
x_fe_master(regs, var_labels = NULL, var_indicates = NULL,
  var_omits = NULL, star_levels = star_level_default, beta_digits,
  se_digits)
}
\arguments{
\item{regs}{stored regression output in a list}

\item{var_labels}{vector of variable labels}

\item{var_indicates}{vector of variables to indicate}

\item{var_omits}{vector of variables to omit}

\item{star_levels}{statistical significance stars}
}
\description{
Helper function to construct coefficients
}
\keyword{internal}
