% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estout.R
\name{estout}
\alias{estout}
\title{Output Regression Results as a LaTeX Table}
\usage{
estout(regs, file = "", var_labels = NULL, var_omits = NULL,
  var_indicates = NULL, sumstat_include = sumstat_include_default,
  star_levels = star_level_default)
}
\arguments{
\item{regs}{stored regression output in a list}

\item{file}{output location, if blank prints to scree}

\item{var_labels}{tibble of variable labels, this sets the order of variables in the table}

\item{var_omits}{vector of variables to omit}

\item{var_indicates}{tibble of variables to indicate, this sets of the order of indicator variables in the table}

\item{sumstat_include}{vector of summary statistics to include.
The defaults are c("N", "aR2", "Ymean")}

\item{star_levels}{vector of cut offs for statistical significance stars.
The defaults are c(0.10, 0.05, 0.01).
Make this 0 for no stars at all}
}
\description{
Like stata's `estout`, output regression results nicely.
}
\examples{

# using lm, glm, and felm with `mtcars`
reg1 <- mtcars \%>\% lfe::felm(data = ., mpg ~ wt | cyl)
reg2 <- mtcars \%>\% lm(data = ., mpg ~ hp)
reg3 <- mtcars \%>\% dplyr::filter(gear == 4) \%>\%
    glm(data = ., vs ~ wt + am, family = binomial(link = "logit"))
reg4 <- mtcars \%>\% lm(data = ., mpg ~ wt + hp + am + as.factor(cyl))
reg5 <- mtcars \%>\% lm(data = ., wt ~ hp)
reg6 <- mtcars \%>\% lfe::felm(data = ., wt ~ hp | cyl + am)

# function will take regressions as a list of list objects
regs <- list(reg1, reg2, reg3, reg4, reg5, reg6)

# and a tibble of variable labels (optional)
var_labels <- tibble::tibble(term = c("wt", "hp"), label = c("Weight", "Horsepower"))
# when labelling an instrumental variable in felm syntax
# the term should be "`x(fit)`" including the `s
# if x is instrumented for by some z

# which variables to omit? (optional)
var_omits <- c("(Intercept)")

# which variables to indicate yes no (good for FEs) (optional)
var_indicates <- tibble::tibble(term = c("am", "cyl"),
    indicator = c("Transmission FE", "Cylinders FE"))

# which summary stats to include?
sumstat_include <- c("N", "aR2", "Ymean")

# estout(regs, file = "", var_labels, var_omits, var_indicates, sumstat_include)

}
