% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat_master.R
\name{sumstat_master}
\alias{sumstat_master}
\title{Gather Summary Statistics from Regression Results for Output}
\usage{
sumstat_master(regs, sumstat_include = "N", sumstat_names)
}
\arguments{
\item{regs}{stored regression output in a list}

\item{sumstat_include}{vector of summary statistics to include}

\item{sumstat_names}{tibble from the lookup with summary stats and proper names}
}
\description{
Helper function to construct summary statistics
}
\keyword{internal}
