% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\docType{data}
\name{sumstat_names}
\alias{sumstat_names}
\title{estout lookups}
\format{An object of class \code{data.frame} with 6 rows and 3 columns.}
\usage{
sumstat_names
}
\arguments{
\item{NULL}{}
}
\description{
Helper function to define various lookups
}
\keyword{internal}
