% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffs.R
\name{x_fe_master}
\alias{x_fe_master}
\title{Compile Regression Coefficients for Output}
\usage{
x_fe_master(regs, var_labels = NULL, var_indicates = NULL,
  var_omits = NULL)
}
\arguments{
\item{regs}{stored regression output in a list}

\item{var_labels}{tibble of variable labels}

\item{var_indicates}{tibble of variables to indicate}

\item{var_omits}{vector of variables to omit}
}
\description{
Helper function to construct coefficients
}
\keyword{internal}
